Adds spec for file buffer

+++ spec/file_buffer_spec.rb
@@ -14,44 +14,75 @@ describe Backchat::WebSocket::FileBuffer do
 
     it "creates the path if missing" do
       buff = FileBuffer.new(work_path)
+      buff.on(:open) do
+        File.exist?(work_path).should be_true
+      end
       buff.open
-      File.exist?(work_path).should be_true
       FileUtils.rm_rf(work_path)
     end
   end
 
   context 'when not draining' do
-    work_path = "./test-work2"
+    work_path = "./rb-test-work2"
     log_path = "#{work_path}/buffer.log"
     exp1 = "the first message"
     exp2 = "the second message"
 
-    it "writes to a file" do
+    before(:each) do
       FileUtils.rm_rf(work_path) if File.exist?(work_path)
+    end
+
+    after(:each) do 
+      FileUtils.rm_rf(work_path)
+    end
+
+    it "writes to a file" do      
       buff = FileBuffer.new(log_path)
       lines = []
-      buff.on(:open) do         
-        buff.write(exp1)
-        buff.write(exp2) do
-          buff.close
-        end
-      end
-      buff.on(:close) do 
-        lines = File.open(log_path).readlines
-        FileUtils.rm_rf(work_path)
-      end
       buff.open
-      lines.size.should == 2
+      buff.write(exp1)
+      buff.write(exp2)
+      buff.close
+      File.open(log_path).readlines.size.should == 2      
     end
   end 
 
   context "when draining" do
+    work_path = "./rb-test-work3"
+    log_path = "#{work_path}/buffer.log"
+    exp1 = "the first message"
+    exp2 = "the second message"
+
+    before(:each) do
+      FileUtils.rm_rf(work_path) if File.exist?(work_path)
+    end
+
+    after(:each) do 
+      FileUtils.rm_rf(work_path)
+    end
+
     it "writes new sends to the memory buffer" do
-      
+      mem = []
+      buff = FileBuffer.new(log_path, :memory_buffer => mem, :initial_state => 1)
+      buff.write(exp1)
+      buff.write(exp2)
+      buff.close
+      mem.size.should == 2     
     end
 
     it "raises data events for every line in the buffer" do
-      
+      lines = []
+      buff = FileBuffer.new(log_path)
+      buff.on(:data) do |args|
+        lines << args
+      end
+      buff.open
+      buff.write(exp1)
+      buff.write(exp2) do
+        buff.drain
+      end
+      buff.close
+      lines.size.should == 2
     end
   end
 end
